fix(text-field): Fixing extra_attributes knobs - FRONT-948 (#363)
